folder_restructure: move all folder containing es6 js files into src folder so that it will be transpiled correctly in prod environment
