+++ event-store/src/app.js
@@ -2,7 +2,7 @@ import grpc from "grpc";
 import { loadSync } from "@grpc/proto-loader";
 import commonGrpcConfig from "../../grpc-common-config";
 import Routes from "./routes";
-import sc from "../clients/nats-streaming";
+import sc from "./clients/nats-streaming";
 
 const packageDef = loadSync("event/event_store.proto", commonGrpcConfig);
 const { EventStoreService } = grpc.loadPackageDefinition(packageDef).event;
