+++ payment/src/app.js
@@ -1,5 +1,5 @@
-import sc from "../clients/nats-streaming";
-import NATSStreamingWorker from "../workers/nats-streaming";
+import sc from "./clients/nats-streaming";
+import NATSStreamingWorker from "./workers/nats-streaming";
 
 sc.on("connect", ()=> {
 	const natsWorker = new NATSStreamingWorker(sc);
